Alter extension key to use i18n for SpecialPage link
The change of the SpecialPage key will lead to the
usage of the corresponding l18n message key for the
link on SpecialPage:SpecialPages.

The name of the SpecialPage key is not that important,
so so decided on changing the key rather than introducing
a new message.

The SpecialPage message key was used since this is the
text of the title on the SpecialPage, is already translated
and fits best for the link title.

Also fixed array syntax to short.

Change-Id: I8a44d38bfbaa5fcee447d5b53203d77549cc553d