Handle duplicate titles gracefully
This patch checks to see if the target title already exists on the
current wiki. If it does it prompts the user to change the title. NOJS.

It also allows the user to click the edit title button to select a new
title (even when there is no conflict)

One TODO has been left in place in SpecialImportFile...
 - Once the name has been changed, check the name doesn't exist on the
   source wiki

Bug: T160160
Change-Id: I318bb0297396523d64253d6046d8c25c06df38c2