Introduce SourceSite interface & finish title checks
For title checks this follows up to
I318bb0297396523d64253d6046d8c25c06df38c2
Upon a rename we now check back on the source site to
see if the new title is allowed.
This is used for mediawiki imports as we dont want to import
to a new title that was used on the source site, as then using
the image on the source site in the case of a remote repo steup
would be annoying.

For review of the title checking the following files should be checked:
 - Interfaces/ImportTitleChecker
 - MediaWiki/ApiImportTitleChecker
 - MediaWiki/HttpApiLookup (added caching logic)

Everything else is basically refactoring around the change

Bug: T160160
Change-Id: I1874000fb76b74b98d262d80c6f1e7bc2f549a53