Check hash existance before title
In the case that an exact file copy already exists on the wiki
we don't want the use to start thinking of a new title for it
when on submit they will be shown another error message that
the file already exists.

So instead run the hash check first and thus avoid the error chain.

Change-Id: I367d02ff1f014c7db26c11322fe10f1435f2018d