Add ImportPlanFactory
ImportPlans are always constructed by the factory now.
This code can now be shared between the special page and a
future API.
The factory also deals with the validation of the ImportPlan
and as it is the only way Plans should be created we can assume
an ImportPlans are valid already.

Change-Id: Idd7daa25778ff821c3b1c4e28b40c83c039780f3