Display error when core would change the title
For titles that include a / or a # core can
automatically do things to the title without
necesarily throwing any sort of error.
In cases that we detect the name being changed
by core in the validator throw an exception
with the detials of the change that would have
been made.

Change-Id: Iee32e7859463b35e2cc83de9979474269a3dec5a