More logging
This adds lots of logging in various places of various levels
for the FileImporter channel.
It also switched from the optional logger interfaces to requiring
one in the constructor for many services.

Each individual import operation now has logging, although
logging within the import done in the WikiRevision class will
go to a different log channel..

It would be great to be able to inject our own log channel here?
or somehow get the logs out for inclusion in our own log channel.

Bug: T180491
Change-Id: I9c5f6d977c62b17c3ef3d92d0d4580e2c1e9fbb3