Track number of users not allowed to access special page
This tracks the number of users that try to access the special page
that are either blocked on the wiki, or have a global block, or
that do not have the correct permissions on the local wiki.

We don't bother to count READONLY or UPLOADDISABLED here as these are
not per user cases but instead just general wiki configuration.

Bug: T187538
Change-Id: Ib513ef1ab70d2ade246f0863ff78c6bdee4b176b