Add missing @covers tags to integration tests
In both cases I do have the impression the tests are meant as integration
tests (even if only one is named like this). The two classes I'm adding
here as @covers tags look like they are extracted from their base class
for the sake of readability, but could as well be inlined. This is
especially notable because both classes have a single public method.
Writing separate tests for these single-method classes is possible, but
probably not worth it.

Change-Id: I44a54d44e1d07c01d02d6582b67b715890273375