Remove isParsable and related unused code from SourceUrl
I found this is unused code that can be removed without any effect.
Notice how the constructor was throwing an exception if the URL was
not parsable. It was impossible to create a SourceUrl object with a
URL that is not parsable. No method could ever return false.

I checked the callers of these methods and found that none checked for
a false being returned. Which is fine, because this could never happen
anyway.

Change-Id: I6669df72417cab334dbed276344f56a03d1f3d1e