Add WikiTextEditor to ChangeFileInfoFormTest coverage
During todays story time I realized this class I listed in T190290 as
not being covered is actually part of ChangeFileInfoForm and part of it's
test.

This does not mean the coverage is sufficent. A browser test is needed to
make sure the editor can actually be used during the process.

Bug: T190290
Change-Id: I10d740bc76408534239f1ef55186b682b6e8de9e