Add missing WikiPage object to edit hook context
By using the DerivativeContext the fallback for the getWikiPage
method does not work and so the page has to be set manualy.

To be more consistent here WikiFilePage is used with a basic sanity
check. This is nothing we would expect to happen though.

Also classes were renamed to better represent the usage.

Bug: T191026
Change-Id: I0270687daf0e89885e58e9b368de2ccee57183de