Make ApiDetailRetrieverTest independend from configuration
Multiple test cases are not totally independend from the default
configuration. By moving the setMwGlobals() call up to the setUp()
method, the same (mocked) configuration values are active for all
tests.

Note how this also makes one of the tests less expensive, because no
100 elements arrays have to be created any more.

This also adds names to the individual test cases in the dataProviders.

Bug: T190821
Change-Id: Icc24126fde03bcfd17a12af32b08b9f8d164e065