Add test for FileImporter\Data\ImportDetails
This is basically just a very trivial value class. But I found the
hashing feature crucial and a lot of test cases to make sure the hash
really changes with every single field.

Note to reviewer: Please have a look at the imageDisplayUrl field! Do
you think it is correct for this field to not be part of the hash
calculation?

Bug: T190290
Change-Id: I1c7420bbe1b0655467d4690090041edb498e4a85