Remove unused parameter defaults and unused method parameters
The idea is to make the code less suprising. If a parameter is optional
(especially if it's the last one in a very long list), it is very hard
to read calls to this method or constructor. Is the last parameter the
optional one, or was it intentionally left out?

Change-Id: I61da2694ec9a2214b99892835bf8280543f2165b