Add @covers tags to cover all trivial Html (snippet) classes
I checked again and I believe all classes in this Html directory are not
worth separate unit tests. The classes do nothing but wrapping a bit of
(literally) HTML. If this works as intended should always be tested via a
higher-level test.

I'm aware these @cover tags create "unwanted" coverage in a few cases.
That is, the code is sometimes called, but there is no specific assertion
that really makes sure the HTML is outputted ecactly as expected. I
believe this is acceptable.

Change-Id: I39820eb600ee0876288fbba35ee7a50b81c6666d