Replace (abandoned) ApexTheme with BlankTheme
The change from WikimediaUITheme to ApexTheme was a mistake, based on
the fact that the file ApexTheme.php appears more blank than the file
BlankTheme.php. But Apex is so blank because it is abandoned, while
Blank should be used in tests. See I6a3acb4 for a longer explanation.

Change-Id: I70d82bd0e8cc0204db429a72befb3e47a8ace5a3