Add test for DuplicateFileRevisionChecker
We are currently not tracking this in a separate Phabricator ticket.
But it is still listed as one of the bigger remaining project risks at
https://doc.wikimedia.org/cover-extensions/FileImporter/dashboard.html

I also narrowed one interface down from LocalRepo to the base class
FileRepo, because the only method called on this object is declared
in the base class.

Bug: T190290
Change-Id: I6f699220288f8ab1309a8a98faaae06498873f78