Fix title validating tests
Using assertSame here shows that one of the tests here went wrong.
The validator did not return a bad file type in that case.
assertEquals compared "true" with "9" here and the test was green.
Although no bad error status was returned.

Looking into the validator, bad title strings will be normalized
and do not result in a bad file type. Only a bad file extension
will trigger that error.

Change-Id: I70f120bcb65de766a9bb0109eaecc0b754b7e187