Fix incomplete error reporting in HttpApiLookup
I discovered something odd here: Note that the HttpRequestException
never sets a message. getMessage can never contain anything for this
particular exception.

Instead, I'm focussing on the message in the StatusValue object.

This patch also makes I3d8342f obsolete. The CURL error is covered by
the very similar code path with the wfMessage() call in this patch.

An easy way to test this is with any invalid URL.

Bug: T196161
Change-Id: If66d8f378567699fb91779676096162b15b6cdc3