Provide non-empty plain text message in HttpRequestException
I found this quite confusing. An exception should always have an
explanatory message, and easily can in this case.

My first approach was to use the message key from the first error in
the StatucValue object. But since this can be empty and the StatusValue
class already contains code for exactly what is needed here, I reuse
what StatusValue::__toString does.

Bug: T196161
Change-Id: I3375741e9b7ba3d83274207ab33fe3f8ff8dc6ff