Docs for Wikimedia vs. 3rd-party MediaWikis vs. non-MediaWikis
We have code that is supposed to work with any source, including
non-MediaWikis, e.g. SourceUrl.

We have code for any MediaWiki, no matter if Wikimedia or 3rd-party.

And we have code that only makes sense on MediaWiki wikis in the
Wikimedia cluster.

Namespaces and documentation should make this more clear.

Change-Id: I152a2b49ee0bd8de78a621be08840b87998165a6