Inline SourceUrl normalization into the SourceSite service
As this patch is now it does not change any behavior. It just moves
existing code around. The goal is to make the way all the services
work together more explicit and easier to understand.

Before: Normalization was a seperate step that was called before the
other steps. But this was optional, not enforced by anything.

Change-Id: Icad215b15400519c16305450a1ca16b1610abd06
After: Normalization is enforced in the two steps.