Prepare full interwiki to prefix map in InterwikiTablePrefixLookup
I'm not sure if this is worth it. This does have two consequences:

1. The loop is guaranteed to run only once, even if multiple sources
are involved. The bad news is: usually there is only one source. So
this optimization probably does not have an effect.

2. The strpos() substring search is reoplaced by an actual comparison
of the (parsed) host name. The downside here is: A few hundred
wfParseUrl() must be done, but probably only one is going to be used.
Is this worth it anyway?

Change-Id: I5d6b9dd0f618e3ab43057b4312acb9a8282b6657