Fix minor code style issues detected via PHPCS
I run the Wikibase CodeSniffer on the FileImporter codebase, and found
a few minor issues this patch fixes. One was detected by the
"FunctionSpacing" sniff, which I'm adding here. Other issues have been
detected by the custom "ClassLevelDocumentation" sniff, which I can not
add at the moment without making this codebase depend on a Wikibase
codebase, which I think should not be done.

Change-Id: I2afb5a7b91c9cd5521ea30f0ee1d4c7ca094f488