Make SiteTableSiteLookup fail on duplicate domains
The sites table does not guarantee the domain is unique. As far as I can
see this is not critical on our production servers, but can be for 3rd
parties. For now I will not invest more time in this, e.g. *not* try to find
the site by comparing more elements from the "link path". But the service at
least needs to stop returning ambiguous results.

Change-Id: I49fcae36edd0a2aa83c00bea9aed713556ff6f71