Make $logger constructor parameter optional
The codebase currently follows two different schemes: In some cases the
logger is a constructor parameter, in other cases there is a setLogger()
method. The later means the logger is a NullLogger by default, until the
setter is called. In tests it does not make sense to call
…->setLogger( new NullLogger() ) as this is the default already.

I would like to streamline these patterns a bit, and this is a first
attempt to do so.

Change-Id: I0d35c4a4bb08c5dd9c769c6e348f25f7a411bc47