Rename "Local > Commons" to more generic "Source > Target"
The word "Commons" is misleading in many situations because most of this
code is not limited to Wikimedia Commons, but can work with any target
wiki.

The word "Local" is a bit confusing, because all this code is running on
the target wiki, technically making Wikimedia Commons the "local"
machine.

Change-Id: I0a5a467dd7dd1fe93daa8060c13eabf173468985