Clean-up and fix bad validation in CommonsHelperConfigParser
One of the "=== false" comparisons was missing in CommonsHelperConfigParser.
Because of this the error message was also shown when the section was not
missing, but empty.

I rewrote the test to expect specific error messages to cover this better.

This patch also removes unused code, and updates some documentation.

All this is done in preparation for T198711.

Change-Id: Icc2e3af02d029136000328588f2fd683d1274330