Switch to manifest_version 2 for extension.json
This also changes the type from "other" to "specialpage".

All described at
https://www.mediawiki.org/wiki/Manual:Extension.json/Schema
The documentation says "Unless you need to use one of the features
implemented in v2, there's no reason to upgrade away from v1."
My reasoning is that the config section was messy and became
uneccessaryly hard to maintain.

Change-Id: Iceb811dc86ae76e3f95458ed7284e0e7cc9e1804