Remove unused code from ImportPlan and add tests
The property was forgotten. It's never set to anything any more, and
always contains null.

The getInitialFileInfoText method was doomed to fail when there is no
latest TextRevision. This is not relevant in production, as there can
not be a page with no revision (or can it?).

This patch also adds trivial assertions for two previously untested
methods, and renames the test case to "testTextGetters" to match what
it now does.

Change-Id: Iea7e3ae08799584bbed45f6bd36725b72daca67d