Do not bind against Message class where not necessary
wfMessage() is the unified, common way to create Message objects. It's
not forbidden to call the Message constructor directly, but there is not
really a benefit in doing so.

In many cases it was not necessary to wrap the message specifier into
a "new Message()" call. What MediaWiki core calls a "message specifier"
can be a string (the message key), a Message object, or an array, where
the first element is the message key and all other elements are the
message parameters. This patch favors the later format.

Change-Id: Iaf23279ebba921d2d1f0f83caf0eea8ec2f3d0d9