Avoid using the real User class in tests, where possible
The additional request mock fixes an unexpected crash when this test runs
in an environment with the TwoColConflict extension enabled. The OutputPage
objects ends in a hook handler that quits without doing anything, just as
expected. But it does a brief check on a request parameter, and this was
missing before.

Bug: T209942
Change-Id: I628ca02bf87f9b4bd98b77d6ce29178a5c573ae1