Integration test for import log entry
This adds a basic integration test if the log entry was created
correctly and is associated with the correct revision.

Since the test already uses the db this seemed a good way to
make sure the entry is added.

The method used to access the test database was customized from
the DatabaseLogEntry::newFromId.

Bug: T209942
Change-Id: I3cb3bdbaabf138b1e1c82a6ac346babeb20d3875