Add "fileimporter" tag to new import log entry and null revision
This is modelled after code I found in core's MovePage. There the tag is
added to the revision first, *before* the log entry is created.

Notes:

[DONE] It seems there is no good way to *query* tags. The tests ended being
somewhat awkward because of this.

[DONE] I wasn't sure if the tag must be somehow "registered" first. Most
code in ChangeTags says it should *not* be used from extensions. It turned
out calling ChangeTags::addTags() is fine.

[DONE] Technically optional, but for the sake of completeness we should
hook into ListDefinedTags to make our tag appear on Special:Tags.

[DONE] I wondered if ChangeTags::addTags is able to create the tag in a way
that it becomes active for the null revision and the log entry the same
time. If this is the case, the $logEntry->setTags call might not be
necesarry, and ChangeTags::addTags can be called with both the $revId
and the $logId. The result should be the same. Turns out it doesn't work
this way, but even simpler.

Bug: T198324
Change-Id: Id07f7f97a1ebadaafbf833493ca4c43803e8a522