Fix unpredictable history order by enforcing >=1 second difference
It's >=1 second and not exactly 1 second because there is no way to
read, modify, or even predict the timestamp of the later edit
(WikiPage::doEditContent does not allow anything). We assume it will
happen the same second, or possibly 1 or more seconds later. The only
guarantee we can give is that the earlier null edit is 1 second in the
past.

This fixes the issue described in T200259. Since seconds are usually not
visible anywhere, it should not have any consequences for the user.

Bug: T200259
Change-Id: I1532283c2ee2cb8064f320032628886bd245eacd