Refactor HTML builders accessing original WebRequest
The issue with this code can be seen when doing a search for "->getVal("
in this codebase. The only place reading user input from the WebRequest
is the central SpecialImportFile page. But in addition to that some of
the HTML builders also access the WebRequest object to read another
value.

This patch moves this additional "importDetailsHash" value into the
ImportRequest object that already contains all the other values
provided via the WebRequest.

This allows further refactoring of these HTML builders.

Bug: T212199
Change-Id: Ifc06729d87eb4a8f6280a415f29b98684371cfda