Add "never create log entries" assertions to ImporterComponentTest
This is a direct follow up for the addition done in I93b9374. The code
added is currently not used because there is currently no test for
multiple file revisions. Here I'm marking these method with "never
expected to be called" to document this fact. This also makes sure we
don't accidentially call the log creation code when it's not needed.

Change-Id: Ic25238d5cca4d67a50c86b13d21e7aae0646b827