Stop as early as possible when there are 0 file/text revisions
I implemented this sanity check as early as possible, in the two
value objects File… & TextRevision. This means that no code working
with these objects will ever be executed with zero file or text
revisions.

The majority of the changes is done in the ImporterComponentTest.
2 of the 3 test cases worked with 0 file revisions before (but 1
text revision). This does not work any more. I fixed the test cases
and deleted one because it became 100% identical to the 3rd one.

Change-Id: I91504987ffec7d256292bf78dd196a653f0a23f6