Guarantee there is a latest file revision
We can now give this guarantee (almost), because the list of file
revisions can not be empty any more.

This code might still fail when the file revisions are incomplete,
have no timestamp, or all the timestamps are invalid. I added a post
condition to make sure this makes the code fail as early as possible.

Change-Id: Icef493af776c376f89fa5d3d8e2d177c3c1d505b