Adding tests for the null revisions upload log entry
This patch adds a test for the upload log entry associated to the
import null revision. The timestamp was added to the database request
to allow picking the upload log entry for a specific revision. This
works in that case because the timestamps should all be different.

It seems that tags are not added for the upload log entry. This will
be investigated further.

Bug: T210755
Change-Id: Iedce86a9ca8a8e83a003a8bb884e6060b8301425