Remove (effectively) unused return value from NullRevisionCreator
The only reason this is (still) returned is to check if it is false or
not. And then we go on and throw a runtime exception anyway. We can
throw the same exception earlier. This is pretty much exactly what
exceptions are for, I would argue.

Note I'm not blaming anyone here, the contrary. The code was intentionally
returning the revision object before because it was needed to create the
log entries (outside of NullRevisionCreator). This changed with the
previous patch.

Change-Id: I7c27ad0b39247552117f81ae7d0ef8172d811a7a