Add missing @covers tags
The …HtmlFragment class got introduced, but I forgot to add it to any
test. The base class could be added to all tests that are testing one of
the subclasses. To not make this to messy I only added it to a few that
are actually using some of the more advanced features of the base class.

The other two service classes never had coverage, as far as I can see. I
feel it is pointless to write stand-alone unit tests for these as they are
pretty trivial. Being covered via the integration test should be more than
enough.

Bug: T212199
Change-Id: I33eef16772613560b2700a0c05e6519cf21754f3