Add RemoteApiImportTitleCheckerTest and fix misplaced log
Before the error log line was created in all situations where a file
with the same name already exists on the target wiki. But the log
message talks about a missing "pages" key.

This mistake was introduced in the original prototype phase, see
I9c5f6d9.

The test makes sure the log call only happens when there is really
something wrong with the API response.

Bug: T180491
Change-Id: Ifd3eb882f5f3a9acea1d9e0e3215ec9e8532cf15