Better message for "section on config page missing"
I found the current message does have 2 issues:
* What do these <> brackets mean?
* What does "about" mean?

We know better. We know this message is only shown if a section is
missing. So why not let the message say this. Even better: Asking the
user to add the section can't be wrong, as this is literally a 100%
copy-paste operation. Critical is the *content* of the section, which
must be adapted to the local wiki, but this is not what this error
message is about.

Change-Id: I476a8e1c896f31d387f941c42ec012a7a5f8818a