Fixup incomplete test coverage in various places
This fixes the code coverage for some tiny bits of code that often
really don't need any test, or got forgotten for one or the other
reason.

* setLogger() methods are not really relevant production code, but more
debug helpers than anything. No point in explicitely covering these.

* The NullRevisionCreator expects a MutableRevisionRecord and does
something special with it. This was introduced after the test have been
written.

* The SiteTableSiteLookup does caching. This was not tested at all
before.

* The code in SiteWiring.php is so trivial that I feel it's actually
counterproductive to try writing a test for it. (If even possible. I
think there is no way to cover code that is not in a class.)

Change-Id: Ife280967e289ee923ae43b11035de5cdbed6b95d