Allow to use subst:… in template transfer rules
This is not really a new feature, but more a bugfix. Before, the
replacement would have been {{subst|…}}. Now, the replacement is simply
what the users ask for, as specified on the configuration page. Excluding
the : character was a mistake. In the ;: wiki syntax it's important to
exclude the : in the first part, but not in the second.

Bug: T214365
Change-Id: Ibf6819b045a45a42d65cfb0350e6d2a565cbd6d9