Show some sensible default when no help page is configured
The default configuration for the help page is empty. If this is
forgotten, the error message is broken: "There is no configuration for
the wiki at ." The message just ends with nothing.

Before I007c32e, we constructed the link. This was a problem and
intentionally replaced with the help page link. Because of this we can
not construct a default. But pointing to at least the server name is at
least better than nothing.

Change-Id: Ib3f43a86e294223e49389e59c95bbbe1d2d126c7