Don't hard-code assumed setup of CommonsHelper config server
Before, the assumption was that *all* wikis would have the same
configuration in terms of script path and article path.

With this change applied, the assumption is that only two wikis have
the same configuration: The wiki running the FileImporter extension,
and the wiki holding the CommonsHelper config pages. It's very likely
these two are actually the same server. However, on the current
Wikimedia setup these servers are not the same but Commons and
mediawiki.org. But we know these two have the same setup.

So even if it is still not nice to make these assumptions, I think
this is already an improvement.

I needed to touch this because I was unable to run local tests with
my setup that does not have a /w/ path.

Change-Id: Ib15a6ff7c9bc9c557aefc71e232f56b4ff0e1217