Simplify error message about incomplete config pages
We changed this just recently via I476a8e1. Both the message before
and after patch I476a8e1 had the same problem: To know if a file is
legally allowed on Commons, only the license stuff is needed. Other
sections are not that critical.

We still want to block all imports as long as the config page is not
complete. We don't want to implement anything about "legally required"
in our code. So the same message is used for all sections. It should
be neutral enough to fit to all of them.

This was discussed and approved by the COM (communication) person
responsible for user facing messages in our codebases.

Change-Id: I9feac748f6072dfbad4a0b580601aaddaf1f1f86