Use prefixed external usernames when importing file revisions
Due to some changes in core it seems we can use the prefixed external
usernames when importing file revisions now.

This also fixes the object expected as a performer on log entries. Due to
changes made in Ic483d0fda931810ac639b9de1cc5e5a4ef8161cc this would not
work otherwise.

The tests for the WikiRevisionFactory were changed to avoid that the user
created with the hook test interferes with other test where it is expected
that the user does not exist.

There seems to be an inconsistency though with the normalization of the
usernames when importing text or file revisions. See T221235

Bug: T220713
Change-Id: I406e4e38eeaae484843ebe91fbb6a01fcfb16f2e