Resolve circular dependency in SpecialImportFile
I used the tool
https://gerrit.wikimedia.org/r/#/admin/projects/mediawiki/tools/dependency-analysis
to find this. It is suprisingly easy: the circular dependency between the
two classes is only because of the constants. The special page uses the
HTML class, but the HTML class also needs the constants. Putting the
comnstants in the special page was simply a mistake, I would argue.

Change-Id: I09ea6960ffcb2da35ef71ea56258c13f9af1e27d