Avoid expensive WebRequest::getVal() where possible
getVal() is expensive because it does Unicode normalization. This is not
needed when we know the parameter contains plain ASCII only, or when it
is only used as a boolean flag. The worst thing that can happen is that
getRawVal() returns a string with invalid UTF-8 sequences, but that's fine
when it is immediately discarded.

Change-Id: I0bb84820b258b5d2eec9ba76845af285aa2deb12