Use read-only end() instead of array_pop() with side-effect
array_pop() is well readable, but it does have a critical side-effect:
it manipulates the source array and actually removes the element. This is
never needed in these cases, as far as I can see. It potentially makes
the following code brittle. It might assume the array element is still
present, but array_pop() removed it.

end() also does have a side-effect: it sets the internal array pointer to
the end. This is usually never a problem as we are never using it (e.g. via
each(), which will be removed from PHP).

This patch also:
* Removes a possibly misplaced limit from a removeNamespaceFromString()
  helper. Links like "en:File:Foo" should be returned as "Foo".
* Replaces some empty() with strict string comparisons. Note that empty()
  considers the string "0" empty.

Change-Id: I9a362b9005a00d618c615cc33722db191ba43b39