Introduce WikiLinkCleaner to be used instead of callables
This allows to have the code that actually processes the wikilinks not
in (possibly unnamed) functions, but in easy to test classes.

I used functions before to have the most trivial, TDD-driven patch.

The next step is to introduce two of these cleaners for each task, and
finally wire everything together.

Bug: T198584
Bug: T213821
Change-Id: I7a6d6b82dc6e4abfdc8f66a919d93d4c99addf0b