Fix heading regex failing on extra whitespace at EOL
As pointed out in I34a6b38.

This patch also makes use of "possessive quantifiers" – the extra plus
at the end of ++, *+, or ?+. This makes sure the quantifier really eats
everything and does not give it back – it stops backtracking. This is
done for performance reasons, and avoids ugly edge-cases like "== =="
possibly resulting in a level 1 heading with the text "= =".

Change-Id: Ib204657b33f3e8db5d35462ac4b87e13f0f97ec6