Integration tests are not code coverage
This is a monarch among integration tests, but we shouldn't consider it as
proper coverage.  For example, we're only verifying the correctness of one
line of ImportPreviewPage, that a particular <h2> element is emitted, yet
the entire file had been marked as 100% covered.  Integration tests are a great
way to find gross typos and get a quick sanity check, but the overly-optimistic
coverage annotations are obscuring reality when we want to write finer-grained
unit tests.

Change-Id: I160db6078f270dff24e43fc5c7ac2fa657443c33