Avoid a few empty() as well as isset()
Some of these are just not necessary, for example when a variable is
guanarteed to be an array. Checking for `=== []` is more strict, type
safe, and equally readable, I would argue.

Some of them can be replaced with ??, which is some nice PHP 7
syntactic sugar that internally still does an isset(). See
http://maettig.com/1546432500 for a longer explanation of ??.

empty() does have some hidden semantics that can become a problem. It
is effectively the same as writing `isset( $var ) && $var == false`. This
contains two potential issues:

1. Typos in the variable name are not reported. The point of isset() is
to check if a variable was declared or not. In case we know a variable
was declared, neither isset() nor empty() should be used.

2. All falsy values are considered false, including empty strings, as
well as the string "0". But when we know a variable can only be an array,
the only falsy case we care about is the empty array.

empty() should be avoided because of this, unless this is really what the
code should do.

Change-Id: I8dd621103e2e623aeeee0d06dd2a43ce6f3b6322