Add WikitextContentCleaner test for multiple source parameters
The key for this to work is the `(array)` cast in
WikitextConversions::getTemplateParameters(), as well as
WikitextContentCleaner::renameTemplateParameters() expecting an array.

A test for the feature in WikitextConversions already exists, see
WikitextConversionsTest::provideTemplateParameterReplacements().

This kind of makes this test an integration test for the two classes
mentioned. Ha, nice. Pure unit test don't exist, right?

Bug: T213955
Change-Id: Icf26e44fe72ec670bf2e2786f12af3325ae0468e