Introduce MediaWikiSourceUrlNormalizer
This fixes an older FIXME in WikimediaSourceUrlNormalizer. Before, the
class encoded normalization rules that are only true for wikis in the
Wikimedia cluster, as well as some that are true for all MediaWiki
installations. With this change these two groups of normalizations are
now split.

Note that the current ServiceWiring intentionally uses
WikimediaSourceUrlNormalizer for both setups, even for the one that
"allows importing files form ANY mediawiki site". This is because we
are using this setup for all our tests, and want the Wikimedia specific
normalizations to happen.

Change-Id: I09aa838523192b8ef277a6cc5e1da64597df4593