Extract duplicate code to MediaWikiSourceUrlParser trait
I'm not sure if a trait is the best method. But it allows for a relatively
small, easy to understand patch. Extracting this into an actual service
that then needs to be injected via all constructors would be way more
complicated.

Another alternative is a static, globally accessible method. But I really
don't like these.

I also tried to come up with a MediaWikiSourceUrl subclass. The biggest
issue: First, the specific type of a SourceUrl is unknown. Only after it
passed one of the SourceUrlChecker it can be converted into a known type.
Some services can then check if they got a MediaWikiSourceUrl and act
accordingly. But that (doing instanceof checks in random places) is one
of the worst bits of subclassing. There is barely any place in the codebase
that can be type-hinted against such a MediaWikiSourceUrl. So no, not
helpful with the current architecture.

Change-Id: I0fc0a0c6427377320dd735a86047e14d68404c09