Change HttpRequestExecutor to accept a parameter array
This patch does have these main effects:

1. There is only a single wfAppendQuery() call in this codebase now.
Also there is no manual URL concatenation with `. '?' .` any more.

2. I had to change a few exception messages as well as log calls. They
can not contain the full API URL any more. Instead, exceptions now only
show the base URL of the API. Logs contain the "apiUrl" and "apiParameters"
in separate fields, or only one of them, depending on what's available.
I would like to argue this is acceptable. Some of the exceptions will be
shown as nice error messages to the users, others won't. In neither case
is a full API URL really helpful. Definitely not for the user. We
developers can restore the API call that is being made in other ways, in
the worst case by looking at the code that triggered an exception/log.

Do you think this is fine?

Change-Id: I645d65d8dc8dcc712f748fde97b92bc60f69e882