Make Media: namespace behave identical to File: namespace
[[Media:…]] is a sudo namespace. As far as I get it the main reason is to
be able to link to files without getting thumbnail images. The same can
be done with [[:File:…]].

In the current state of this parser all File links are excluded, no matter
if prefixed with a colon or not. For consistency Media links should behave
the same.

This behavior might change later depending on the outcome of T223290.

Bug: T198584
Change-Id: I29a8f96094cb02ac251823f11e4b5cf0659233ee