Use empty TextContent instead of mock in tests
The TextContent class is relatively simple. Replacing it with a
mock does not provide much of a benefit. It's especially not faster.

This also does make the test setup more compatible with code in
other extensions that might be triggered via hooks, and expect the
content object to do something specific. This is not relevant any
more in our CI setup after we disabled all hooks (see I1342314),
but does make it easier to work on this test code in the future.

Change-Id: I1c4db88726198fec58063d71ebb82b3b34c43a90