Make CommonsHelperConfigRetriever an injectable service
The main change of this refactoring is the removal of the SourceUrl
parameter from the constructor. Before, it was not possible to inject
this service, because the SourceUrl was expected on construction time.

This refactoring is intentionally "incomplete" and leaves another FIXME
in the code. This is to be done in later patches.

Change-Id: I86f8e53b10064d28e3fdf00e9f1f6e094fbe6d5e