Move knowledge about the API back to ApiDetailRetriever
I would like to do this for these reasons:

1. Conceptually, knowledge about the structure of these arrays of arrays
and the "ns" and "title" keys is knowledge about how the API works. It's
not a lot and certainly not a big issue. But still.

2. I plan to continue moving the wikitext validation and cleanup steps
out of the ApiDetailRetriever. To be able to do this, I must store the
lists of templates and categories somewhere. I would like to minimize the
memory footprint for this, and only store two flat arrays of strings with
irrelevant entries already removed.

Change-Id: Idb77f1db1e71b797bad135edbd1c0c06cae3f689